refactor: update the export service's request
After talking to Cody from the Export Service, we have agreed that what
applications receive are actually resource requests and not export
requests.

An export request is generated by the user and is received by the export
service, which then sends multiple resource requests to the Kafka
channel, which in turn are picked up by the corresponding applications.
Then, the export service is responsible for gathering all the responses
from the applications and building a proper response to that export request.

Therefore, technically speaking, the applications just receive resource
requests associated to an export request.